Signed/unsigned comparison
git-svn-id: https://uberdemotools.googlecode.com/svn/trunk@32 605c6810-b171-2bb6-8537-90259a8e6946

+++ UDT_DV/src/qt/demo_player.cpp
@@ -223,7 +223,7 @@ void DemoPlayer::updateEntityList(int startIndex, int time)
 	}
 
 	// Update Players
-	for(int i = 0; i < demo->_playerPlaybackInfos.size(); i++)
+	for(size_t i = 0; i < demo->_playerPlaybackInfos.size(); i++)
 	{
 		const Demo::PlayerInfo& info = demo->_playerPlaybackInfos[i];
 		if(info.Time == serverTime)
